update build scripts

+++ web-app/scripts/install_port.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 echo "installing UShER port..."
-if [ "$1" == "latest" ]; then
-	"downloading latest port."
+if [[ "$1" == "latest" ]]; then
+	echo "downloading latest port."
 
 	remote_file=$(curl -s https://api.github.com/repos/amkram/shusher/releases/latest \
 	| grep "browser_download_url.*gz" \
@@ -16,7 +16,12 @@ if [ "$1" == "latest" ]; then
 	rm $local_file
 
 else
-	"custom path to port provided: $1"
-	cp $1/usher.* dist/
+	echo "custom path to port provided: $1"
+	if [[ -f $1/usher.wasm && -f $1/usher.js && -f $1/usher.data &&  -f $1/usher.worker.js ]]; then
+		cp $1/usher.* dist/
+		echo "done installing port."
+	else
+		echo "ERROR: usher files not found in $1. Please check your path"
+
+	fi
 fi
-echo "done installing port."
